+++ Mastodon/Scene/SuggestionAccount/TableView-Components/SuggestionAccountTableViewCell.swift
@@ -82,30 +82,18 @@ final class SuggestionAccountTableViewCell: UITableViewCell {
     func configure(viewModel: SuggestionAccountTableViewCell.ViewModel) {
         userView.configure(user: viewModel.user, delegate: delegate)
 
-        Publishers.CombineLatest3(
-            viewModel.followedUsers,
-            viewModel.followRequestedUsers,
-            viewModel.blockedUsers
-        )
-        .receive(on: DispatchQueue.main)
-        .sink { [weak self] followed, requested, blocked in
-            
-            guard let self else { return }
-            
-            if blocked.contains(viewModel.user.id) {
-                self.userView.setButtonState(.blocked)
-            } else if followed.contains(viewModel.user.id) {
-                self.userView.setButtonState(.unfollow)
-            } else if requested.contains(viewModel.user.id) {
-                self.userView.setButtonState(.pending)
-            } else if viewModel.user.locked {
-                self.userView.setButtonState(.request)
-            } else {
-                self.userView.setButtonState(.follow)
-            }
+        if viewModel.blockedUsers.contains(viewModel.user.id) {
+            self.userView.setButtonState(.blocked)
+        } else if viewModel.followedUsers.contains(viewModel.user.id) {
+            self.userView.setButtonState(.unfollow)
+        } else if viewModel.followRequestedUsers.contains(viewModel.user.id) {
+            self.userView.setButtonState(.pending)
+        } else if viewModel.user.locked {
+            self.userView.setButtonState(.request)
+        } else {
+            self.userView.setButtonState(.follow)
         }
-        .store(in: &disposeBag)
-        
+
         let metaContent: MetaContent = {
             do {
                 let mastodonContent = MastodonContent(content: viewModel.user.note ?? "", emojis: viewModel.user.emojis.asDictionary)
